Made multipiliers carry over and stack with consecutive splits as long as the player clears something with each split:
- This should drastically improve gameplay as it encourages players to plan higher scoring splits over multiple moves rather than try and survive or build a single turn

+++ Quick Split/Assets/Scripts/Game Screen/Piece Related/piece_script.cs
@@ -29,7 +29,7 @@ public class piece_script : MonoBehaviour {
 	public int groupValue;
 
 	//stores multiplier to reflect accurate score;
-	public int multiplier;
+	public float multiplier;
 
 	public GameController gameController;
 
@@ -304,7 +304,7 @@ public class piece_script : MonoBehaviour {
 		Vector2 spawnPoint = GameObject.FindGameObjectWithTag ("MainCamera").GetComponent<Camera>().WorldToViewportPoint (transform.position);
 		int scoreValue;
 		if (multiplier != 0)
-			scoreValue = groupValue * multiplier;
+			scoreValue = (int)(groupValue * multiplier);
 		else
 			scoreValue = groupValue;
